Update 02.in_silico_digestion_PstIGal7.R
Dear Violeta,

I wanted to bring to your attention a small issue we have encountered with our window boundaries. The current script assumes the presence of a motif at the beginning and end of each chromosome, but it seems that the last windows of each chromosome are being extended by 1 bp. To ensure accurate window annotation, I have added a command to check for this issue and correct it by subtracting 1 bp. Please rest assured that this will not affect any of your work.

Thank you for your attention to this matter.